requirements.txt
pyJWT==1.4.2
djangorestframework-jwt==1.8.0
djangorestframework==3.3
-r requirements/staging.txt